+++ lib/uvc_camera_controller.dart
@@ -111,9 +111,11 @@ class UVCCameraController {
 
   /// 获取当前摄像头请求参数
   Future<String?> getCurrentCameraRequestParameters() async {
-    return await _cameraChannel?.invokeMethod('getCurrentCameraRequestParameters');
+    return await _cameraChannel
+        ?.invokeMethod('getCurrentCameraRequestParameters');
   }
 
+  /// 更新预览大小
   void updateResolution(PreviewSize? previewSize) {
     _cameraChannel?.invokeMethod('updateResolution', previewSize?.toMap());
   }