+++ lib/uvc_camera_view.dart
@@ -29,7 +29,12 @@ class UVCCameraViewParamsEntity {
   });
 
   Map<String, dynamic> toMap() {
-    return {"minFps": minFps, "maxFps": maxFps, "frameFormat": frameFormat, "bandwidthFactor": bandwidthFactor};
+    return {
+      "minFps": minFps,
+      "maxFps": maxFps,
+      "frameFormat": frameFormat,
+      "bandwidthFactor": bandwidthFactor
+    };
   }
 }
 
@@ -39,7 +44,11 @@ class UVCCameraView extends StatefulWidget {
   final double height;
   final UVCCameraViewParamsEntity? params;
   const UVCCameraView(
-      {super.key, required this.cameraController, required this.width, required this.height, this.params});
+      {super.key,
+      required this.cameraController,
+      required this.width,
+      required this.height,
+      this.params});
 
   @override
   State<UVCCameraView> createState() => _UVCCameraViewState();