+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 
 
 
+## 0.0.3
+
+- feat getAllPreviewSizes updateResolution 
+- feat set Camera Preview Parameters
+- fix some issues
+
 ## 0.0.2
 
 - update documentation